requirements.txt
shioaji
pandas
pyyaml
loguru
pytest
